.gpg was missing from the local path

+++ start.sh
@@ -26,9 +26,9 @@ fi
 while true; do
   find ${SHARE_DIR} -name "${FILE_NAME}" | while read LINE; do
     gpg --recipient "${GPG_UID}" --encrypt "${LINE}"
-    rm -f "${LINE}"
-    aws s3 cp --region ${AWS_REGION} "${LINE}" "s3://${AWS_BUCKET}/$(basename ${LINE}).gpg"
-    rm -f "$LINE.gpg"
+    rm -f "${LINE:?}"
+    aws s3 cp --region ${AWS_REGION} "${LINE:?}.gpg" "s3://${AWS_BUCKET}/$(basename ${LINE}).gpg"
+    rm -f "${LINE:?}.gpg"
     echo "UPLOADED: ${LINE}"
   done;
   sleep 10;
